feat[#365] : 스낵바 overlay.tsx -> snackbar.tsx 로 변경
